+++ packages/perseus-core/src/data-schema.ts
@@ -1675,7 +1675,7 @@ export type PerseusNumericInputWidgetOptions = {
      * Translatable Text; Text to describe this input. This will be shown to
      * users using screenreaders.
      */
-    // TODO(LEMS-4354): Make labelText required and default it to "" in the
+    // TODO(benchristel): Make labelText required and default it to "" in the
     //  parser.
     labelText?: string | undefined;
     /**
@@ -1730,6 +1730,7 @@ export type PerseusNumberLineWidgetOptions = {
      * The position of the endpoints of the number line. Setting the range
      * constrains the position of the answer and the labels.
      */
+    // TODO(benchristel): type `range` as `[number, number]`.
     range: number[];
     /**
      * This controls the position of the left / right labels. By default, the
@@ -2260,7 +2261,10 @@ export type PerseusIFrameWidgetOptions = {
     height: number | string;
     /** Whether to allow the IFrame to become full-screen (like a video) */
     allowFullScreen: boolean;
-    /** Whether to allow the iframe content to redirect the page */
+    /**
+     * Whether to allow the iframe content to redirect the page
+     * @deprecated - not used in Perseus.
+     */
     allowTopNavigation?: boolean;
     /** Always false */
     static: boolean;