+++ packages/perseus/src/widgets/orderer/orderer.tsx
@@ -17,12 +17,10 @@ import type {
 } from "../../types";
 import type {OrdererPromptJSON} from "../../widget-ai-utils/orderer/orderer-ai-utils";
 import type {
-    PerseusOrdererWidgetOptions,
     PerseusOrdererUserInput,
     OrdererPublicWidgetOptions,
 } from "@khanacademy/perseus-core";
 import type {LinterContextProps} from "@khanacademy/perseus-linter";
-import type {PropsFor} from "@khanacademy/wonder-blocks-core";
 
 type PlaceholderCardProps = {
     width: number | null | undefined;
@@ -320,18 +318,6 @@ type OrdererState = {
     onAnimationEnd?: (arg1: any) => void;
 };
 
-// eslint-disable-next-line no-restricted-syntax
-0 as any as WidgetProps<
-    PerseusOrdererWidgetOptions,
-    PerseusOrdererUserInput
-> satisfies PropsFor<typeof WrappedOrderer>;
-
-// eslint-disable-next-line no-restricted-syntax
-0 as any as WidgetProps<
-    OrdererPublicWidgetOptions,
-    PerseusOrdererUserInput
-> satisfies PropsFor<typeof WrappedOrderer>;
-
 class Orderer
     extends React.Component<OrdererProps, OrdererState>
     implements Widget
@@ -382,7 +368,7 @@ class Orderer
             // @ts-expect-error - TS2322 - Type 'number' is not assignable to type 'null'.
             placeholderIndex = index;
         } else {
-            content = this.props.options[index].content;
+            content = this.props.options.options[index].content;
         }
 
         this.props.handleUserInput({current: list});
@@ -449,7 +435,7 @@ class Orderer
         if (inCardBank) {
             // If we're in the card bank, go through the options to find the
             // one with the same content
-            this.props.options.forEach((opt, i) => {
+            this.props.options.options.forEach((opt, i) => {
                 if (opt.content === this.state.dragContent) {
                     const card = ReactDOM.findDOMNode(this.refs["bank" + i]);
                     // @ts-expect-error - TS2769 - No overload matches this call. | TS2339 - Property 'position' does not exist on type 'JQueryStatic'.
@@ -505,7 +491,7 @@ class Orderer
 
     findCorrectIndex: (arg1: any, arg2: any) => any = (draggable, list) => {
         // Find the correct index for a card given the current cards.
-        const isHorizontal = this.props.layout === "horizontal";
+        const isHorizontal = this.props.options.layout === "horizontal";
         // @ts-expect-error - TS2769 - No overload matches this call.
         const $dragList = $(ReactDOM.findDOMNode(this.refs.dragList));
         // @ts-expect-error - TS2339 - Property 'offset' does not exist on type 'JQueryStatic'.
@@ -552,7 +538,7 @@ class Orderer
             return false;
         }
 
-        const isHorizontal = this.props.layout === "horizontal";
+        const isHorizontal = this.props.options.layout === "horizontal";
         // @ts-expect-error - TS2769 - No overload matches this call.
         const $draggable = $(ReactDOM.findDOMNode(draggable));
         // @ts-expect-error - TS2769 - No overload matches this call.
@@ -598,8 +584,9 @@ class Orderer
      * [LEMS-3185] do not trust serializedState
      */
     getSerializedState(): any {
-        const {userInput, ...rest} = this.props;
+        const {userInput, options, ...rest} = this.props;
         return {
+            ...options,
             ...rest,
             current: userInput.current.map((e) => ({content: e})),
         };
@@ -685,7 +672,7 @@ class Orderer
         // This is the bank of stacks of cards
         const bank = (
             <div ref="bank" className="bank perseus-clearfix">
-                {this.props.options.map((opt, i) => {
+                {this.props.options.options.map((opt, i) => {
                     return (
                         <Card
                             ref={"bank" + i}
@@ -712,10 +699,10 @@ class Orderer
                 className={
                     "draggy-boxy-thing orderer " +
                     "height-" +
-                    this.props.height +
+                    this.props.options.height +
                     " " +
                     "layout-" +
-                    this.props.layout +
+                    this.props.options.layout +
                     " " +
                     "blank-background " +
                     "perseus-clearfix "